Fix conflicts v4.6.6 with master

See merge request desktop/desktop!1300

+++ src/MEGASync/gui/WaitingSpinnerWidget.cpp
@@ -245,7 +245,7 @@ void WaitingSpinnerWidget::updateSize() {
 }
 
 void WaitingSpinnerWidget::updateTimer() {
-    _timer->setInterval(1000 / (_numberOfLines * _revolutionsPerSecond));
+    _timer->setInterval(static_cast<int>(1000 / (_numberOfLines * _revolutionsPerSecond)));
 }
 
 void WaitingSpinnerWidget::updatePosition() {
